# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "A Future Corporation API",
        "version": "1.0.0"
    },
    "paths": {
        "/payment/nab": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Pay with NAB",
                "description": "Pay credit card with NAB",
                "operationId": "payNab",
                "requestBody": {
                    "description": "Credit card details",
                    "required": true,
                    "content": {
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment succeeded.",
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Payment failed response",
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/payment/anz": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Pay with ANZ",
                "description": "Pay credit card with ANZ",
                "operationId": "payAnz",
                "requestBody": {
                    "description": "Credit card details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment succeeded.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Payment failed response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreditCard": {
                "description": "Credit Card",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CreditCardNumberOnly"
                    },
                    {
                        "properties": {
                            "card_name": {
                                "title": "Name",
                                "description": "Credit card name",
                                "type": "string",
                                "example": "John Doe"
                            },
                            "card_number": {
                                "title": "Number",
                                "description": "Credit card number",
                                "type": "integer",
                                "format": "int64",
                                "example": 374245455400126
                            },
                            "valid_until": {
                                "title": "Valid Until",
                                "description": "Credit card valid until",
                                "type": "string",
                                "format": "date",
                                "example": "2022-02-22"
                            },
                            "ccv": {
                                "title": "CCV number",
                                "description": "Credit card CCV number",
                                "type": "integer",
                                "format": "int32",
                                "example": 123
                            }
                        }
                    }
                ]
            },
            "CreditCardNumberOnly": {
                "description": "Credit Card Number Only",
                "properties": {
                    "card_number": {
                        "title": "Number",
                        "description": "Credit card number",
                        "type": "integer",
                        "format": "int64",
                        "example": 374245455400126
                    }
                },
                "type": "object"
            },
            "Payment": {
                "description": "Payment",
                "properties": {
                    "from": {
                        "$ref": "#/components/schemas/CreditCard"
                    },
                    "merchant_id": {
                        "title": "Merchant Id",
                        "description": "ID of merchant",
                        "type": "integer",
                        "format": "int32",
                        "example": 123
                    },
                    "merchant_key": {
                        "title": "Merchant Key",
                        "description": "Key of merchant",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "amount": {
                        "title": "Amount to be paid",
                        "description": "Amount to be paid",
                        "type": "number",
                        "format": "float",
                        "example": 14.2
                    }
                },
                "type": "object"
            },
            "PaymentResponse": {
                "description": "Payment Response",
                "properties": {
                    "from": {
                        "$ref": "#/components/schemas/CreditCardNumberOnly"
                    },
                    "transaction_number": {
                        "title": "Transaction number",
                        "description": "Number of transaction",
                        "type": "integer",
                        "format": "int32",
                        "example": 123
                    },
                    "transaction_time": {
                        "title": "Transaction time",
                        "description": "Time of transaction",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-02-02 02:02:02"
                    },
                    "amount": {
                        "title": "Amount paid",
                        "description": "Amount paid",
                        "type": "number",
                        "format": "float",
                        "example": 14.2
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}